add seperate example config test
